html/semantics/embedded-content/media-elements/track/track-element/add-text-track.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8" />
	<link rel="help" href="http://www.w3.org/html/wg/drafts/html/CR/embedded-content-0.html#text-track-api" />
    <title>track test</title>	
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
</head>
<body>
	<h1>addTextTrack Method Support & verificate setted properties Test</h1>
	<p>This test will check whether the browser support <em>addTextTrack Method</em> </p>
	<p>
		And, verificate that setted properties are equal to used arguments for addTextTrack Methos.
	</p>
	<video controls loop id="media" style="display:none;">
		<source src="http://media.w3.org/2010/05/sintel/trailer.mp4" type="video/mp4" id="media">
	</video>
	<div id="log"></div>
	<script type="text/javascript">
		test(function() {
			var vod = document.getElementById("media");
			assert_not_equals(vod.addTextTrack, undefined ,"support addTextTrack Method");
		}, "support addTextTrack Method");

		test(function() {			
			var vod = document.getElementById("media");
			var txtTrck = vod.addTextTrack("subtitles", "korean", "ko");
			assert_equals(txtTrck.kind, "subtitles" , "kind value Equal test");
			assert_equals(txtTrck.label, "korean" , "label value Equal test");
			assert_equals(txtTrck.language, "ko" , "lang value Equal test");
		}, "properties correct test");

		

	</script>
</body>
</html>
